chore: moving statement_parser into utils folder
